+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/database/transformers/NotificationTransform.java
@@ -1,4 +1,4 @@
-package edu.kit.ipd.crowdcontrol.objectservice.database.transforms;
+package edu.kit.ipd.crowdcontrol.objectservice.database.transformers;
 
 import edu.kit.ipd.crowdcontrol.objectservice.database.model.tables.records.NotificationRecord;
 import edu.kit.ipd.crowdcontrol.objectservice.proto.Notification;
